batch: Move semantic comparison into package
The line-matching package provides alignment while semantic change
comparison remains at the batch package root.

Replacement selection and batch filtering derive their change runs from
matching, so leaving comparison elsewhere obscures that direct dependency.

This commit moves semantic comparison beside the matching algorithm and
keeps the selected import path available during coverage migration.

The remaining line-sequence, range-view, and lineage modules will complete
the package contents.
